% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultMeta.R
\name{bayes_inference}
\alias{bayes_inference}
\title{Summary statistics from a posterior distribution}
\usage{
bayes_inference(x, alp)
}
\arguments{
\item{x}{Univariate sample from the posterior distribution of a parameter.}

\item{alp}{Significance level used in the computation of the credible interval}
}
\value{
a matrix with summary statistics
}
\description{
Given a univariate sample drawn from the posterior distribution, this
function computes the posterior mean, the posterior median, the posterior
standard deviation, and the limits of the \eqn{(1-\alpha)} probability-symmetric
credible interval.
}
