% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH_sample_post.R
\name{duplication_matrix}
\alias{duplication_matrix}
\title{Duplication matrix}
\usage{
duplication_matrix(p)
}
\arguments{
\item{p}{Integer which specifies the dimension of the duplication matrix.}
}
\value{
a matrix of size \eqn{p^2 \times p(p+1)/2}
}
\description{
This function creates the duplication matrix of size \eqn{p^2 \times p(p+1)/2}
}
