% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultMeta.R
\name{plot.BayesMultMeta}
\alias{plot.BayesMultMeta}
\title{Plot a BayesMultMeta object}
\usage{
\method{plot}{BayesMultMeta}(x, ...)
}
\arguments{
\item{x}{a BayesMultMeta object}

\item{...}{additional arguments}
}
\value{
No return value, produces trace plots
}
\description{
This function produces the trace plots of the constructed Markov chains.
}
