% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMultMeta.R
\name{summary.BayesMultMeta}
\alias{summary.BayesMultMeta}
\title{Summary statistics from the posterior of a BayesMultMeta class}
\usage{
\method{summary}{BayesMultMeta}(object, alpha = 0.95, ...)
}
\arguments{
\item{object}{BayesMultMeta class}

\item{alpha}{Significance level used in the computation of the credible interval.}

\item{...}{not used}
}
\value{
a list with summary statistics
}
\description{
Summary statistics from the posterior of a BayesMultMeta class
}
